fix: dont crash during request handlers
